pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "piri"
version = "1.0.0"
description = "Lightweight LLM + RAG Engine by AKIS Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AKIS Platform", email = "dev@akisplatform.com"},
]
keywords = ["llm", "rag", "ai", "nlp", "faiss", "turkish", "akis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.6.0",
    "faiss-cpu>=1.7.4",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "datasets>=2.18.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/akis-platform/piri"
Documentation = "https://github.com/akis-platform/piri#readme"
Repository = "https://github.com/akis-platform/piri"
Issues = "https://github.com/akis-platform/piri/issues"

[tool.setuptools.packages.find]
include = ["piri*"]
